pyproject.toml
[project]
name = "ttr-bot"
version = "0.1.0"
description = "Toontown Rewritten automation bot"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "pyautogui>=0.9.54",
    "pillow>=10.2.0",
    "pyobjc-framework-quartz>=10.1",
    "pyobjc-framework-cocoa>=10.1",
    "pynput>=1.7.6",
    "pytesseract>=0.3.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttr_bot"]

[project.scripts]
ttr-bot = "ttr_bot.__main__:main"

[dependency-groups]
dev = []
